Refactor widget.js into TypeScript modules
Split the monolithic `widget.js` into focused, single-responsibility
TypeScript files, leaving only the Jupyter/AMD factory in widget.ts.

The previous JS monolith was previously useful because the JupyterLab
builds required webpack (which doesn't support TS out-out-of-the-box).
But we switched to rspack, and since these modules _require_ bundling,
it makes sense to change them to TS because the syntax is nicer and
type-stripping is a "free" default with our build tools.


```sh
Before                          After
──────────────────────────────────────────────────────
src/
└── widget.js  (741 lines)      src/
                                ├── widget.ts      ( 98 lines)  ← Jupyter/AMD factory only
                                ├── runtime.ts     (113 lines)
                                ├── binding.ts     (128 lines)
                                ├── host.ts        ( 68 lines)
                                ├── load.ts        (114 lines)
                                ├── invoke.ts      ( 40 lines)
                                ├── observe.ts     ( 16 lines)
                                ├── model-proxy.ts ( 34 lines)
                                ├── widget-ref.ts  (  8 lines)
                                └── util.ts        ( 63 lines)
```
