+++ packages/anywidget/__tests__/widget.test.ts
@@ -1,7 +1,7 @@
 import * as widgets from "@jupyter-widgets/base";
 import * as baseManager from "@jupyter-widgets/base-manager";
-import { afterEach, expect, it } from "vite-plus/test";
-import { page, userEvent } from "vite-plus/test/browser";
+import { afterEach, expect, it } from "vitest";
+import { page, userEvent } from "vitest/browser";
 
 import create_anywidget from "../src/widget.js";
