+++ packages/anywidget/src/host.ts
@@ -1,0 +1,63 @@
+import type { Host } from "@anywidget/types";
+import type { DOMWidgetModel, DOMWidgetView } from "@jupyter-widgets/base";
+
+import { BINDINGS } from "./binding.ts";
+import { invoke } from "./invoke.ts";
+import { model_proxy } from "./model-proxy.ts";
+import { parse_widget_ref } from "./widget-ref.ts";
+
+export function create_host(model: DOMWidgetModel, { signal }: { signal: AbortSignal }): Host {
+  let host: Host = {
+    // @ts-expect-error - model_proxy returns AnyModel; generic T is erased at runtime
+    async getModel(ref) {
+      let model_id = parse_widget_ref(ref);
+      let child_model = await model.widget_manager.get_model(model_id);
+      let context = Symbol("anywidget.host.getModel");
+      signal.addEventListener("abort", () => child_model.off(null, null, context));
+      return model_proxy(child_model, context);
+    },
+    // @ts-expect-error - generic T is erased at runtime, exports typed as unknown
+    async getWidget(ref) {
+      let model_id = parse_widget_ref(ref);
+      let child_model = await model.widget_manager.get_model(model_id);
+      let child_binding = BINDINGS.get(child_model);
+      if (!child_binding) {
+        throw new Error(`[anywidget] No binding found for widget ${model_id}`);
+      }
+      let exports = await Promise.race([
+        child_binding.ready,
+        new Promise((_, reject) =>
+          AbortSignal.timeout(10000).addEventListener("abort", () =>
+            reject(new Error(`[anywidget] Timed out waiting for widget ${model_id} to initialize`)),
+          ),
+        ),
+      ]);
+      return {
+        exports,
+        async render({ el, signal: view_signal }) {
+          let child_view_signal = view_signal ?? signal;
+          // Create a minimal view-like object for the binding
+          // oxlint-disable-next-line typescript-eslint/no-unsafe-type-assertion -- intentionally creating a partial DOMWidgetView for child rendering
+          let child_view = {
+            model: child_model,
+            el,
+            $el: {
+              empty() {
+                el.innerHTML = "";
+              },
+            },
+          } as unknown as DOMWidgetView;
+          await child_binding.create_view(child_view, {
+            signal: child_view_signal,
+            experimental: {
+              // @ts-expect-error - bind isn't working
+              invoke: invoke.bind(null, child_model),
+            },
+            host,
+          });
+        },
+      };
+    },
+  };
+  return host;
+}
