+++ packages/anywidget/src/index.js
@@ -1,4 +1,4 @@
-import create from "./widget.js";
+import create from "./widget.ts";
 
 // @ts-expect-error -- define is a global provided by the notebook runtime.
 define(["@jupyter-widgets/base"], create);
