+++ packages/anywidget/src/widget-ref.ts
@@ -1,0 +1,8 @@
+export let WIDGET_REF_PREFIX = "anywidget:";
+
+export function parse_widget_ref(ref: unknown): string {
+  if (typeof ref === "string" && ref.startsWith(WIDGET_REF_PREFIX)) {
+    return ref.slice(WIDGET_REF_PREFIX.length);
+  }
+  throw new Error(`[anywidget] Invalid widget reference: ${JSON.stringify(ref)}`);
+}
